+++ core/common/servent.cpp
@@ -2579,7 +2579,7 @@ void Servent::sendRawMetaChannel(int interval)
                                     url.convertTo(String::T_META);
 
                                     sprintf(tmp, "StreamTitle='%s';StreamUrl='%s';\0", title.cstr(), url.cstr());
-                                    int len = ((strlen(tmp) + 15+1) / 16);
+                                    int len = (((int)strlen(tmp) + 15+1) / 16);
                                     sock->writeChar(len);
                                     sock->write(tmp, len*16);
